requirements.txt
Flask
gunicorn 
encord
entsoe-py
pandas
python-dotenv
oauthlib
requests 
requests_oauthlib
sentinelhub
